Create Module Tracking url and Carrier List and create two news vendors
