+++ Mag2/Del01CarrierList/RapidCarrierList/etc/frontend/routes.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0"?>
+<!--
+/**
+ * DISCLAIMER
+ *
+ * @category  CarrierList
+ * @package   CarrierList\RapidCarrierList
+ * @author    Del01
+ * @copyright 2021 Del01
+ * @license   Open Software License ("OSL") v. 3.0
+ */
+ -->
+<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+        xsi:noNamespaceSchemaLocation="urn:magento:framework:App/etc/routes.xsd">
+    <router id="standard">
+        <route id="del01carrierlist_rapidcarrierlist" frontName="activecarriers">
+            <module name="Del01CarrierList_RapidCarrierList" />
+        </route>
+    </router>
+</config>
