+++ Mag2/Del01CarrierList/RapidCarrierList/registration.php
@@ -1,0 +1,19 @@
+<?php
+
+/**
+ * DISCLAIMER
+ *
+ * @category  Del01CarrierList
+ * @package   Del01CarrierList\TrackingUrl
+ * @author    Del01
+ * @copyright 2021 del01
+ * @license   Open Software License ("OSL") v. 3.0
+ */
+
+use Magento\Framework\Component\ComponentRegistrar;
+
+ComponentRegistrar::register(
+    ComponentRegistrar::MODULE,
+    'Del01CarrierList_RapidCarrierList',
+    __DIR__
+);
