+++ Mag2/Del01TrackingUrl/TrackingUrl/etc/acl.xml
@@ -1,0 +1,25 @@
+<?xml version="1.0"?>
+<!-- DISCLAIMER
+ * @category  Del01TrackingUrl
+ * @package   Del01TrackingUrl\TrackingUrl
+ * @author    del01
+ * @copyright 2021 del01
+ * @license   Open Software License ("OSL") v. 3.0
+ */
+ -->
+<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+        xsi:noNamespaceSchemaLocation="urn:magento:framework:Acl/etc/acl.xsd">
+    <acl>
+        <resources>
+            <resource id="Magento_Backend::admin">
+                <resource id="Magento_Backend::stores">
+                    <resource id="Magento_Backend::stores_settings">
+                        <resource id="Magento_Config::config">
+                            <resource id="Del01_TrackingLink::tracking" title="Tracking Settings" translate="title" sortOrder="6"/>
+                        </resource>
+                    </resource>
+                </resource>
+            </resource>
+        </resources>
+    </acl>
+</config>
