+++ Mag2/Del01TrackingUrl/TrackingUrl/etc/adminhtml/system.xml
@@ -1,0 +1,33 @@
+<?xml version="1.0"?>
+<!-- DISCLAIMER
+ * @category  Del01TrackingUrl
+ * @package   Del01TrackingUrl\TrackingUrl
+ * @author    del01
+ * @copyright 2021 del01
+ * @license   Open Software License ("OSL") v. 3.0
+ */
+ -->
+<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:module:Magento_Config:etc/system_file.xsd">
+    <system>
+        <section id="del01_tracking" translate="label" type="text" sortOrder="309" showInDefault="1" showInWebsite="1" showInStore="0">
+            <label>Tracking Settings</label>
+            <tab>sales</tab>
+            <resource>Del01_TrackingLink::tracking</resource>
+            <group id="service_url" translate="label" type="text" sortOrder="1" showInDefault="1" showInWebsite="1" showInStore="0">
+                <label>Tracking Service Url</label>
+                <field id="dhl" translate="label" type="text" sortOrder="10" showInDefault="1" showInWebsite="1" showInStore="0" canRestore="1">
+                    <label>dhl</label>
+                </field>
+                <field id="fedex" translate="label" type="text" sortOrder="20" showInDefault="1" showInWebsite="1" showInStore="0" canRestore="1">
+                    <label>fedex</label>
+                </field>
+                <field id="ups" translate="label" type="text" sortOrder="30" showInDefault="1" showInWebsite="1" showInStore="0" canRestore="1">
+                    <label>UPS</label>
+                </field>
+                <field id="usps" translate="label" type="text" sortOrder="40" showInDefault="1" showInWebsite="1" showInStore="0" canRestore="1">
+                    <label>USPS</label>
+                </field>
+            </group>
+        </section>
+    </system>
+</config>
