+++ Mag2/Del01TrackingUrl/TrackingUrl/registration.php
@@ -1,0 +1,19 @@
+<?php
+
+/**
+ * DISCLAIMER
+ *
+ * @category  Del01TrackingUrl
+ * @package   Del01TrackingUrl\TrackingUrl
+ * @author    Del01
+ * @copyright 2021 del01
+ * @license   Open Software License ("OSL") v. 3.0
+ */
+
+use Magento\Framework\Component\ComponentRegistrar;
+
+ComponentRegistrar::register(
+    ComponentRegistrar::MODULE,
+    'Del01TrackingUrl_TrackingUrl',
+    __DIR__
+);
