--- Mag2/RapidCarrierList/etc/frontend/routes.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0"?>
-<!--
-/**
- * DISCLAIMER
- *
- * @category  Mag2
- * @package   Mag2\RapidCarrierList
- * @author    Nathan Marcy <nathan.marcy@live.fr>
- * @copyright 2021 nmarcy
- * @license   Open Software License ("OSL") v. 3.0
- */
- -->
-<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-        xsi:noNamespaceSchemaLocation="urn:magento:framework:App/etc/routes.xsd">
-    <router id="standard">
-        <route id="mag2_activecarrierlist" frontName="activecarrier">
-            <module name="Mag2_ActiveCarrierList" />
-        </route>
-    </router>
-</config>
