--- Mag2/RapidCarrierList/etc/module.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0"?>
-<!--
-/**
- * DISCLAIMER
- *
- * @category  Mag2
- * @package   Mag2\RapidCarrierList
- * @author    Nathan Marcy <nathan.marcy@live.fr>
- * @copyright 2021 nmarcy
- * @license   Open Software License ("OSL") v. 3.0
- */
- -->
-<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:Module/etc/module.xsd">
-    <module name="Mag2_ActiveCarrierList" />
-</config>
