--- Mag2/TrackingUrl/view/frontend/templates/tracking/popupNew.phtml
@@ -1,79 +1,0 @@
-<?php
-
-/**
- * DISCLAIMER
- *
- * @category  Mag2
- * @package   Mag2\TrackingUrl
- * @author    Nathan Marcy <nathan.marcy@live.fr>
- * @copyright 2021 nmarcy
- * @license   Open Software License ("OSL") v. 3.0
- */
-
-use Magento\Framework\View\Element\Template;
-
-/** @var $block \Magento\Shipping\Block\Tracking\Popup */
-
-$results = $block->getTrackingInfo();
-$viewModel = $block->getViewModel();
-?>
-<div class="page tracking">
-    <?php if (!empty($results)) : ?>
-        <?php foreach ($results as $shipId => $result) : ?>
-            <?php if ($shipId) : ?>
-                <div class="order subtitle caption"><?= /* @noEscape */ $block->escapeHtml(__('Shipment #')) . $shipId ?></div>
-            <?php endif; ?>
-            <?php if (!empty($result)) : ?>
-                <?php foreach ($result as $counter => $track) : ?>
-                    <div class="table-wrapper">
-                        <?php
-                            $shipmentBlockIdentifier = $shipId . '.' . $counter;
-                            $block->addChild('shipping.tracking.details.' . $shipmentBlockIdentifier, Template::class, [
-                                'track' => $track,
-                                'template' => 'Therascience_Shipping::tracking/details.phtml',
-                                'storeSupportEmail' => $block->getStoreSupportEmail(),
-                                'nativeTrackUrl' => $viewModel->getTrackingUrl($shipId ,$track)
-                            ]);
-                        ?>
-                        <?= /* @noEscape */ $block->getChildHtml('shipping.tracking.details.' . $shipmentBlockIdentifier) ?>
-                    </div>
-                    <?php if (is_object($track) && !empty($track->getProgressdetail())) : ?>
-                        <?php
-                            $block->addChild('shipping.tracking.progress.' . $shipmentBlockIdentifier, Template::class, [
-                                'track' => $track,
-                                'template' => 'Magento_Shipping::tracking/progress.phtml'
-                            ]);
-                        ?>
-                        <?= /* @noEscape */ $block->getChildHtml('shipping.tracking.progress.' . $shipmentBlockIdentifier) ?>
-                    <?php endif; ?>
-                <?php endforeach; ?>
-            <?php else : ?>
-                <div class="message info empty">
-                    <div><?= $block->escapeHtml(__('There is no tracking available for this shipment.')) ?></div>
-                </div>
-            <?php endif; ?>
-        <?php endforeach; ?>
-    <?php else : ?>
-        <div class="message info empty">
-            <div><?= $block->escapeHtml(__('There is no tracking available.')) ?></div>
-        </div>
-    <?php endif; ?>
-    <div class="actions">
-        <button type="button"
-                title="<?= $block->escapeHtmlAttr(__('Close Window')) ?>"
-                class="action close"
-                onclick="window.close(); window.opener.focus();">
-            <span><?= $block->escapeHtml(__('Close Window')) ?></span>
-        </button>
-    </div>
-</div>
-<script>
-    require([
-        'jquery'
-    ], function (jQuery) {
-        /* hide the close button when the content doesn't open in a modal window */
-        if (window.opener === null || typeof window.opener === "undefined") {
-            jQuery('.actions button.close').hide();
-        }
-    });
-</script>
